importsdk: reuse Lightning file rules for Aurora mapping
