+++ pkg/importsdk/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "importsdk",
     srcs = [
-        "aurora.go",
         "config.go",
         "error.go",
         "file_scanner.go",
@@ -31,7 +30,6 @@ go_library(
         "//pkg/parser/mysql",
         "//pkg/planner/core",
         "//pkg/table/tables",
-        "//pkg/util/table-filter",
         "@com_github_pingcap_errors//:errors",
         "@org_uber_go_zap//:zap",
     ],
@@ -41,7 +39,6 @@ go_test(
     name = "importsdk_test",
     timeout = "short",
     srcs = [
-        "aurora_test.go",
         "config_test.go",
         "file_scanner_test.go",
         "job_manager_test.go",
